+++ Changelog.md
@@ -1,8 +1,9 @@
+# Changelog
+
 #### next version
 
-#### to be released
-
 * added the option to **mine asteroids** for the player
+* mining stations buy ore and plutonium ore (for a very low price)
 
 #### v1.1